draft of log-reg dist mode; some refactoring
