+++ src/apply_gather.h
@@ -17,118 +17,6 @@
 #ifndef _APPLY_GATHER_INCLUDED_
 #define _APPLY_GATHER_INCLUDED_
 
-#include "mpi.h"
-
-namespace map_reduce_star {
-
-struct MPI
-{
-    template<typename T>
-    static std::vector<T> gather(int rank, int nRanks, const T& p_res )
-    {
-        // Serialize the partial result into a data archive
-        daal::data_management::InputDataArchive in_arch;
-        p_res->serialize(in_arch);
-
-        // gather all partial results
-        // First get all sizes, then gather on root
-        int sizes[nRanks];
-        int offsets[nRanks];
-        int mysize = in_arch.getSizeOfArchive();
-        MPI_Gather(&mysize, 1, MPI_INT, sizes, 1, MPI_INT, 0, MPI_COMM_WORLD);
-
-        int tot_sz = 0;
-        char * buff = NULL;
-
-        if(rank == 0) {
-            for(int i = 0; i < nRanks; i++) {
-                offsets[i] = i > 0 ? sizes[i-1] : 0;
-                tot_sz += sizes[i];
-            }
-            buff = new char[tot_sz];
-        }
-
-        MPI_Gatherv(in_arch.getArchiveAsArraySharedPtr().get(), mysize, MPI_CHAR,
-                    buff, sizes, offsets, MPI_CHAR,
-                    0, MPI_COMM_WORLD);
-
-        std::vector<T> all;
-        if(rank == 0) {
-            all.resize(nRanks);
-            for(int i=0; i<nRanks; ++i) {
-                // FIXME: This is super-inefficient, we need to write our own DatArchive to avoid extra copy
-                daal::data_management::OutputDataArchive out_arch(reinterpret_cast<daal::byte*>(buff+offsets[i]), sizes[i]);
-                all[i] = daal::services::dynamicPointerCast<typename T::ElementType>(out_arch.getAsSharedPtr());
-            }
-            delete [] buff;
-        }
-
-        return all;
-    }
-
-    template<typename T>
-    static T bcast(int rank, int nRanks, T obj)
-    {
-        if(rank == 0) {
-            // Serialize the partial result into a data archive
-            daal::data_management::InputDataArchive in_arch;
-            obj->serialize(in_arch);
-            int size = in_arch.getSizeOfArchive();
-            MPI_Bcast(&size, 1, MPI_INT, 0, MPI_COMM_WORLD);
-            MPI_Bcast(in_arch.getArchiveAsArraySharedPtr().get(), size, MPI_CHAR, 0, MPI_COMM_WORLD);
-        } else {
-            int size = 0;
-            MPI_Bcast(&size, 1, MPI_INT, 0, MPI_COMM_WORLD);
-            char * buff = new char[size];
-            MPI_Bcast(buff, size, MPI_CHAR, 0, MPI_COMM_WORLD);
-            daal::data_management::OutputDataArchive out_arch(reinterpret_cast<daal::byte*>(buff), size);
-            obj = daal::services::dynamicPointerCast<typename T::ElementType>(out_arch.getAsSharedPtr());
-        }
-        return obj;
-    }
-};
-
-static const daal::data_management::NumericTablePtr get_table(const TableOrFList & t)
-{
-    if(t.table) return t.table;
-    if(t.file.size()) return readCSV(t.file);
-    assert("only one input per process supported" == NULL);
-    return daal::data_management::NumericTablePtr();
-};
-
-template<typename Algo>
-class map_reduce_star
-{
-public:
-    template<typename ...Ts>
-    typename Algo::iomstep2Master_type::result_type
-    static map_reduce(Algo & algo, Ts& ... inputs)
-    {
-        int rank, nRanks;
-        MPI_Comm_size(MPI_COMM_WORLD, &nRanks);
-        MPI_Comm_rank(MPI_COMM_WORLD, &rank);
-
-        auto s1_result = algo.run_step1Local(inputs...);
-        // gather all partial results
-        auto p_results = MPI::gather(rank, nRanks, s1_result);
-        // call reduction on root
-        typename Algo::iomstep2Master_type::result_type res;
-        if(rank == 0) res = algo.run_step2Master(p_results);
-        // bcast final result
-        return MPI::bcast(rank, nRanks, res);
-    }
-
-    template<typename ... Ts>
-    static typename Algo::iomstep2Master_type::result_type
-    compute(Algo & algo, Ts& ... inputs)
-    {
-        return map_reduce(algo, get_table(inputs)...);
-    }
-};
-
-} // namespace map_reduce_star {
-
-
 #include "cnc4daal.h"
 #include "readGraph.h"
 #include <cstdlib>
