+++ src/daal4py.cpp
@@ -464,6 +464,16 @@ TableOrFList::TableOrFList(PyObject * input)
     }
 }
 
+const daal::data_management::NumericTablePtr get_table(const TableOrFList & t)
+{
+    if(t.table) return t.table;
+    if(t.file.size()) return readCSV(t.file);
+    if(tlist.size()==1) return t.tlist[0];
+    if(flist.size()==1) return readCSV(t.flist[0]);
+    throw std::invalid_argument("one and only one input per process supported");
+    return daal::data_management::NumericTablePtr();
+}
+
 const daal::data_management::NumericTablePtr readCSV(const std::string& fname)
 {
     daal::data_management::FileDataSource< daal::data_management::CSVFeatureManager >
